generate_video_demo/requirements.txt
absl-py==0.9.0
astor==0.8.1
astroid==2.4.2
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
gast==0.2.2
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
numpy==1.19.1
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==7.0.0
pip==20.1.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.5.3
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
setuptools==49.2.0.post20200714
six==1.15.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
Werkzeug==1.0.1
wheel==0.34.2
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
